back_up_file

+++ app/src/main/java/com/pixelplex/qtum/ui/fragment/BackupContractsFragment/BackupContractsFragmentPresenter.java
@@ -14,10 +14,12 @@ import com.pixelplex.qtum.datastorage.FileStorageManager;
 import com.pixelplex.qtum.datastorage.TinyDB;
 import com.pixelplex.qtum.datastorage.backupmodel.Backup;
 import com.pixelplex.qtum.datastorage.backupmodel.ContractJSON;
-import com.pixelplex.qtum.datastorage.backupmodel.Template;
+import com.pixelplex.qtum.datastorage.backupmodel.TemplateJSON;
 import com.pixelplex.qtum.datastorage.model.ContractTemplate;
 import com.pixelplex.qtum.ui.activity.MainActivity.MainActivity;
+import com.pixelplex.qtum.ui.fragment.BaseFragment.BaseFragment;
 import com.pixelplex.qtum.ui.fragment.BaseFragment.BaseFragmentPresenterImpl;
+import com.pixelplex.qtum.utils.DateCalculator;
 
 import java.io.File;
 import java.io.FileOutputStream;
@@ -25,6 +27,12 @@ import java.io.IOException;
 import java.io.OutputStreamWriter;
 import java.util.ArrayList;
 import java.util.List;
+import java.util.concurrent.Callable;
+
+import rx.Observable;
+import rx.Observer;
+import rx.android.schedulers.AndroidSchedulers;
+import rx.schedulers.Schedulers;
 
 
 /**
@@ -58,9 +66,9 @@ public class BackupContractsFragmentPresenter extends BaseFragmentPresenterImpl 
     public void onViewCreated() {
         super.onViewCreated();
 
-        if(getView().getMainActivity().checkPermission(Manifest.permission.WRITE_EXTERNAL_STORAGE)){
-            createBackUpFile();
-        }
+//        if(getView().getMainActivity().checkPermission(Manifest.permission.WRITE_EXTERNAL_STORAGE)){
+//            createBackUpFile();
+//        }
 
         getView().getMainActivity().addPermissionResultListener(new MainActivity.PermissionsResultListener() {
             @Override
@@ -77,7 +85,9 @@ public class BackupContractsFragmentPresenter extends BaseFragmentPresenterImpl 
     @Override
     public void onDestroyView() {
         super.onDestroyView();
-        mBackUpFile.delete();
+        if(mBackUpFile!=null) {
+            mBackUpFile.delete();
+        }
         getView().getMainActivity().removePermissionResultListener();
     }
 
@@ -93,11 +103,58 @@ public class BackupContractsFragmentPresenter extends BaseFragmentPresenterImpl 
 
         if(mBackUpFile==null) {
             createBackUpFile();
+        } else {
+            chooseShareMethod();
         }
+    }
+
+    private void createBackUpFile(){
+        getView().setProgressDialog();
+        createBackupData().subscribeOn(Schedulers.io())
+                .observeOn(AndroidSchedulers.mainThread())
+                .subscribe(new Observer<String>() {
+                    @Override
+                    public void onCompleted() {
+
+                    }
+
+                    @Override
+                    public void onError(Throwable e) {
+
+                    }
+
+                    @Override
+                    public void onNext(String backupData) {
+                        String fileName = "qtum_backup_file.json";
+                        File backupFile = new File(Environment.getExternalStorageDirectory(),fileName);
+
+                        try {
+                            FileOutputStream fOut = null;
+                            fOut = new FileOutputStream(backupFile, true);
+                            OutputStreamWriter osw = new OutputStreamWriter(fOut);
+                            osw.write(backupData);
+                            osw.flush();
+                            osw.close();
+                        } catch (IOException e) {
+                            e.printStackTrace();
+                        }
+
+                        String backUpFileSize = String.valueOf((int)Math.ceil(backupFile.length()/1024.0)) + " Kb";
+                        getView().dismissProgressDialog();
+                        getView().setUpFile(backUpFileSize);
+                        mBackUpFile = backupFile;
+
+                        chooseShareMethod();
+                    }
+                });
+
+    }
+
+    public void chooseShareMethod(){
         Intent intentShareFile = new Intent(Intent.ACTION_SEND);
 
         if(mBackUpFile.exists()) {
-            intentShareFile.setType("application/txt");
+            intentShareFile.setType("application/json");
             intentShareFile.putExtra(Intent.EXTRA_STREAM, Uri.parse("file://"+mBackUpFile.getAbsolutePath()));
 
             intentShareFile.putExtra(Intent.EXTRA_SUBJECT,
@@ -107,54 +164,38 @@ public class BackupContractsFragmentPresenter extends BaseFragmentPresenterImpl 
         }
     }
 
-    private void createBackUpFile(){
-        String fileName = "qtum_backup_file.txt";
-        String backupData = createBackupData();
-        File backupFile = new File(Environment.getExternalStorageDirectory(),fileName);
 
-        try {
-            FileOutputStream fOut = null;
-            fOut = new FileOutputStream(backupFile, true);
-            OutputStreamWriter osw = new OutputStreamWriter(fOut);
-            osw.write(backupData);
-            osw.flush();
-            osw.close();
-        } catch (IOException e) {
-            e.printStackTrace();
-        }
+    public Observable<String> createBackupData(){
+        return rx.Observable.fromCallable(new Callable<String>() {
+            @Override
+            public String call() throws Exception {
+                TinyDB tinyDB = new TinyDB(mContext);
 
-        String backUpFileSize = String.valueOf((int)Math.ceil(backupFile.length()/1024.0)) + " Kb";
-        getView().setUpFile(backUpFileSize);
+                List<TemplateJSON> templateJSONList = new ArrayList<>();
 
-        mBackUpFile = backupFile;
-    }
+                List<ContractTemplate> contractTemplateList = tinyDB.getContractTemplateList();
+                for(ContractTemplate contractTemplate : contractTemplateList){
+                    String source = FileStorageManager.getInstance().readSourceContract(mContext, contractTemplate.getUiid());
+                    String bytecode = FileStorageManager.getInstance().readByteCodeContract(mContext,contractTemplate.getUiid());
+                    String abi = FileStorageManager.getInstance().readAbiContract(mContext,contractTemplate.getUiid());
+                    TemplateJSON templateJSON = new TemplateJSON(source,bytecode,contractTemplate.getUiid(),contractTemplate.getDate(),abi,contractTemplate.getContractType(),contractTemplate.getName());
+                    templateJSONList.add(templateJSON);
+                }
 
-    private String createBackupData(){
-        TinyDB tinyDB = new TinyDB(mContext);
+                List<ContractJSON> contractList1 = new ArrayList<>();
 
-        List<Template> templateList = new ArrayList<>();
+                List<Contract> contractList = tinyDB.getContractList();
+                for(Contract contract : contractList){
+                    String contractTemplateType = tinyDB.getContractTemplateByUiid(contract.getUiid()).getContractType();
+                    ContractJSON contract1 = new ContractJSON(contract.getContractName(),contract.getSenderAddress(),contract.getContractAddress(),contractTemplateType,contract.getDate(),contract.getUiid(),contract.isSubscribe());
+                    contractList1.add(contract1);
+                }
 
-        List<ContractTemplate> contractTemplateList = tinyDB.getContractTemplateList();
-        for(ContractTemplate contractTemplate : contractTemplateList){
-            String source = FileStorageManager.getInstance().readSourceContract(mContext, contractTemplate.getUiid());
-            String bytecode = FileStorageManager.getInstance().readByteCodeContract(mContext,contractTemplate.getUiid());
-            String abi = FileStorageManager.getInstance().readAbiContract(mContext,contractTemplate.getUiid());
-            Template template = new Template(source,bytecode,contractTemplate.getUiid(),"test",abi,contractTemplate.getContractType(),contractTemplate.getName());
-            templateList.add(template);
-        }
-
-
-        List<ContractJSON> contractList1 = new ArrayList<>();
-
-        List<Contract> contractList = tinyDB.getContractList();
-        for(Contract contract : contractList){
-            ContractJSON contract1 = new ContractJSON(contract.getContractName(),contract.getSenderAddress(),contract.getContractAddress(),"test","test",contract.getUiid(),false);
-            contractList1.add(contract1);
-        }
-
-        Backup backup = new Backup("test",templateList,"test","test",contractList1,"android");
-        Gson gson = new Gson();
-        return gson.toJson(backup);
+                Backup backup = new Backup(DateCalculator.getCurrentDate(), templateJSONList,String.valueOf(android.os.Build.VERSION.SDK_INT),"test",contractList1,"android");
+                Gson gson = new Gson();
+                return gson.toJson(backup);
+            }
+        });
     }
 
 }
